requirements.txt
pytest==7.3.1
black==23.3.0
aiohttp==3.8.4
asyncio==3.4.3
allure-pytest==2.13.2
async-timeout==4.0.2
ujson==5.7.0
pytest-asyncio==0.21.0
flake8==6.0.0
selenium
pytest-xdist
pytest-rerunfailures
webdriver-manager
requests==2.31.0
